cosine_general: just have one representative A
Don't try to do too much at the same time Conway.
But what if I want to extend it to A other than 3?
Well the way you had it before, you hard-coded the whole list of A values,
  so to extract any individual value of A, e.g. 10,
  you would then need to specify its position in the list.
  E.g. for A = 10 you would have to write aValuesGen[[2]],
  so the magic constant changes from 10 to 2.
All right. Why are we having this internal monologue?
I dunno. Why are you writing it into this commit message?
I don't know either. Bytes are cheap?