FigureStyles.wl: make traced boundaries thicker
so that there is increased contrast against background curves,
which were made darker in aec8adf9b2.
Also amalgamate "TracedUpper" and "TracedLower",
which need not exist since it is clearer to label the branches
as upper and lower (rather than use dotting/dashing,
which is confusing given the styling for contours
and for the terminal curve).